Split the xml and yaml subclasses into their own files.
